it's working

+++ neural_artistic_style_wrapper.py
@@ -2,6 +2,7 @@ NEURAL_ARTISTIC_STYLE_DIR = '/home/daniel/app/neural_artistic_style'
 import os
 import subprocess
 import shutil
+import sys
 
 def run():
 	
@@ -9,10 +10,14 @@ def run():
 		return False
 	os.mknod('.crunching.lck')
 
-	shutil.rmtree(os.path.join('images', 'animation'))
+	style = sys.argv[1]
+	subject = sys.argv[2]
+	
+	if os.path.isdir(os.path.join('images', 'animation')):
+		shutil.rmtree(os.path.join('images', 'animation'))
 	args = ['python', '-u', os.path.join(NEURAL_ARTISTIC_STYLE_DIR, 'neural_artistic_style.py'), \
-		'--style', os.path.join(NEURAL_ARTISTIC_STYLE_DIR, 'images', 'donelli.jpg'), \
-		'--subject', os.path.join(NEURAL_ARTISTIC_STYLE_DIR, 'images', 'weird_al.jpg'), \
+		'--style', style, \
+		'--subject', subject, \
 		'--iterations', '200', \
 		'--vgg19', os.path.join(NEURAL_ARTISTIC_STYLE_DIR, 'imagenet-vgg-verydeep-19.mat'), \
 		'--output', os.path.join('images', 'out.png'), \
